move parsePortMapping() to resources.go
