+++ pkg/spec/populators.go
@@ -21,14 +21,11 @@ import (
 	"fmt"
 	"sort"
 	"strconv"
-	"strings"
 
 	log "github.com/Sirupsen/logrus"
 	api_v1 "k8s.io/client-go/pkg/api/v1"
 
 	"github.com/pkg/errors"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/client-go/pkg/api"
 )
 
 func populateServicePortNames(serviceName string, servicePorts []api_v1.ServicePort) {
@@ -165,93 +162,6 @@ func populateEnvFrom(c Container, cms []ConfigMapMod, secrets []SecretMod) (Cont
 	envs = append(envs, c.Env...)
 	c.Env = envs
 	return c, nil
-}
-
-// Parse the string the get the port, targetPort and protocol
-// information, and then return the resulting ServicePort object
-func parsePortMapping(pm string) (*api_v1.ServicePort, error) {
-
-	// The current syntax for portMapping is - port:targetPort/protocol
-	// The only field mandatory here is "port". There are 4 possible cases here
-	// which are handled in this function.
-
-	// Case 1 - port
-	// Case 2 - port:targetPort
-	// Case 3 - port/protocol
-	// Case 4 - port:targetPort/protocol
-
-	var port int32
-	var targetPort intstr.IntOrString
-	var protocol api_v1.Protocol
-
-	protocolSplit := strings.Split(pm, "/")
-	switch len(protocolSplit) {
-
-	// When no protocol is specified, we set the protocol to TCP
-	// Case 1 - port
-	// Case 2 - port:targetPort
-	case 1:
-		protocol = api_v1.ProtocolTCP
-
-	// When protocol is specified
-	// Case 3 - port/protocol
-	// Case 4 - port:targetPort/protocol
-	case 2:
-		switch api_v1.Protocol(protocolSplit[1]) {
-		case api_v1.ProtocolTCP, api_v1.ProtocolUDP:
-			protocol = api_v1.Protocol(protocolSplit[1])
-		default:
-			return nil, fmt.Errorf("invalid protocol '%v' provided, the acceptable values are '%v' and '%v'", protocolSplit[1], api.ProtocolTCP, api.ProtocolUDP)
-		}
-	// There is no case in which splitting by "/" provides < 1 or > 2 values
-	default:
-		return nil, fmt.Errorf("invalid syntax for protocol '%v' provided, use 'port:targetPort/protocol'", pm)
-	}
-
-	portSplit := strings.Split(pm, ":")
-	switch len(portSplit) {
-
-	// When only port is specified
-	// Case 1 - port
-	// Case 3 - port/protocol
-	case 1:
-		// Ignoring the protocol part, if present, and converting only the port
-		// part
-		p, err := strconv.ParseInt(strings.Split(portSplit[0], "/")[0], 10, 32)
-		if err != nil {
-			return nil, errors.Wrap(err, "port is not an int")
-		}
-
-		port, targetPort.IntVal = int32(p), int32(p)
-
-	// When port and targetPort both are specified
-	// Case 2 - port:targetPort
-	// Case 4 - port:targetPort/protocol
-	case 2:
-		p, err := strconv.ParseInt(portSplit[0], 10, 32)
-		if err != nil {
-			return nil, errors.Wrap(err, "port is not an int")
-		}
-		port = int32(p)
-
-		// Ignoring the protocol part, if present, and converting only the
-		// targetPort part
-		tp, err := strconv.ParseInt(strings.Split(portSplit[1], "/")[0], 10, 32)
-		if err != nil {
-			return nil, errors.Wrap(err, "targetPort is not an int")
-		}
-		targetPort.IntVal = int32(tp)
-
-	// There is no case in which splitting by ": provides < 1 or > 2 values
-	default:
-		return nil, fmt.Errorf("invalid syntax for portMapping '%v', use 'port:targetPort/protocol'", pm)
-	}
-
-	return &api_v1.ServicePort{
-		Port:       port,
-		TargetPort: targetPort,
-		Protocol:   protocol,
-	}, nil
 }
 
 func populateContainers(containers []Container, cms []ConfigMapMod, secrets []SecretMod) ([]api_v1.Container, error) {
